requirements.txt
azure-functions
azure-storage-blob==12.26.0
langchain
langchain-community
langchain-text-splitters
langchain-openai
langchain-chroma
openai
chromadb
tiktoken
